Move scheduling formulations to generic module
